refactor(constants): keep constants.py at core/resources/ [AE-2946]
Reverts the file move from the previous commit per review feedback — too
much churn for a refactor. Keeps the URL centralization, bug fixes, and
new URL constants; they now live in the original
src/runpod_flash/core/resources/constants.py.

- Restore src/runpod_flash/core/resources/constants.py (from core/).
- Restore tests/unit/core/resources/test_constants.py (from tests/unit/core/).
- Revert 21 import sites back to runpod_flash.core.resources.constants
  (absolute) and .constants / from ..constants (relative).
- core/api/runpod.py reads RUNPOD_API_BASE_URL and RUNPOD_REST_API_URL
  inline (duplicated with constants.py defaults) to avoid a circular
  import: core.resources.__init__ eagerly imports modules that re-enter
  core/api/runpod.py, so that module can't import from core.resources.constants.
