+++ tests/unit/core/resources/test_constants.py
@@ -5,7 +5,7 @@ from unittest.mock import patch
 
 import pytest
 
-from runpod_flash.core.constants import (
+from runpod_flash.core.resources.constants import (
     CPU_PYTHON_VERSIONS,
     DEFAULT_PYTHON_VERSION,
     GPU_BASE_IMAGE_PYTHON_VERSION,
