+++ tests/unit/test_endpoint_client.py
@@ -327,7 +327,7 @@ class TestEnsureEndpointReadyIdMode:
         with patch(
             "runpod_flash.endpoint.ENDPOINT_DOMAIN", "api.runpod.ai", create=True
         ):
-            from runpod_flash.core.constants import ENDPOINT_DOMAIN
+            from runpod_flash.core.resources.constants import ENDPOINT_DOMAIN
 
             with patch.object(
                 ep,
